# quell-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Quell AI Backend (derived from routes)
  version: 1.0.0
  description: |
    OpenAPI specification for the Quell AI backend API. Endpoints cover authentication,
    store management, Shopify OAuth, chat webhook, inbox, support tickets, appearance,
    analytics, and admin metrics.
servers:
  - url: https://histioid-imogene-spinally.ngrok-free.dev
    description: Production
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      properties:
        id: { type: integer }
        session_id: { type: string }
        store_url: { type: string }
        channel: { type: string }
        user_message: { type: string, nullable: true }
        bot_response: { type: string, nullable: true }
        message_json: { type: object, nullable: true }
        is_read: { type: integer }
        created_at: { type: string, format: date-time }
...
security:
  - BearerAuth: []
paths:
...
  /api/shopify/install:
    get:
      tags: [Shopify]
      summary: Redirect to Shopify OAuth installation URL
      parameters:
        - in: query
          name: shop
          schema: { type: string }
      responses:
        '302':
          description: Redirect to Shopify
  /api/shopify/callback:
    get:
      tags: [Shopify]
      summary: Shopify OAuth callback that exchanges code and stores pending registration
      parameters:
        - in: query
          name: code
          schema: { type: string }
        - in: query
          name: shop
          schema: { type: string }
        - in: query
          name: hmac
          schema: { type: string }
      responses:
        '302':
          description: Redirects to frontend signup/login
...
  /api/inbox:
    get:
      tags: [Inbox]
      summary: List conversation sessions (grouped) with pagination and filters
      parameters:
        - in: query
          name: store_url
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, default: 30 }
        - in: query
          name: channel
          schema: { type: string }
        - in: query
          name: only_unread
          schema: { type: boolean }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Paginated session list
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      page: { type: integer }
                      perPage: { type: integer }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        session_id: { type: string }
                        channel: { type: string }
                        last_message_at: { type: string }
                        last_user_message: { type: string }
                        last_bot_response: { type: string }
                        unread_count: { type: integer }
  /api/inbox/{sessionId}/messages:
    get:
      tags: [Inbox]
      summary: Fetch full message thread for a session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
        - in: query
          name: store_url
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Thread messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId: { type: string }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Message" }
  /api/inbox/{sessionId}/mark-read:
    post:
      tags: [Inbox]
      summary: Mark a session as read (emits socket update)
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                store_url: { type: string }
      responses:
        '200':
          description: Success
  /api/inbox/{sessionId}/reply:
    post:
      tags: [Inbox]
      summary: Seller reply — save message and emit to client
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                store_url: { type: string }
                message: { type: string }
                channel: { type: string }
                user_email: { type: string }
              required: [store_url, message]
      responses:
        '200':
          description: Message saved (returns messageId)
...